Added animation editor ui file.
