+++ editor/src/core/ui/views/opened_project/windows/details_ui.h
@@ -2,8 +2,23 @@
 
 #include "../../../imgui/imgui_helper.h"
 
+class AnimatedSpriteComp;
+
+// Temp
+struct FuncObject {
+    FuncObject(std::function<void()> func) {
+        func();
+    }
+};
+
 namespace OpenedProjectUI {
 namespace Windows {
 ImGuiHelper::Window GetDetailsWindow();
+
+class AnimationEditor : public Singleton<AnimationEditor> {
+  public:
+    explicit AnimationEditor(singleton) {}
+    ImGuiHelper::PopupModal& GetPopup(AnimatedSpriteComp* animatedSpriteComp);
+};
 } // namespace Windows
 } // namespace OpenedProjectUI
